tweak numbers, move define to single file
