+++ addons/laserpointer/CfgWeapons.hpp
@@ -38,13 +38,7 @@ class CfgWeapons {
         baseWeapon = "acc_pointer_IR";
 
         class ItemInfo: ItemInfo {
-            class Pointer: Pointer {
-                isIR = 0; // visible laser
-                irDotSize = QUOTE(0.1/4); // default size / 4
-                beamThickness = 0; // no visible beam
-                dotColor[] = {8192, 0, 0}; // 5mW/low-vis
-                beamColor[] = {0, 0, 0};
-            };
+            POINTER_VISIBLE_RED;
         };
     };
 
@@ -58,12 +52,7 @@ class CfgWeapons {
         baseWeapon = "ACE_acc_pointer_green";
 
         class ItemInfo: ItemInfo {
-            class Pointer: Pointer { // properties can't be inherited
-                isIR = 0;
-                beamThickness = 0;
-                dotColor[] = {0, 16384, 0}; // high-vis
-                beamColor[] = {0, 0, 0};
-            };
+            POINTER_VISIBLE_GREEN;
         };
     };
 };
