+++ addons/laserpointer/script_component.hpp
@@ -15,3 +15,4 @@
 #endif
 
 #include "\z\ace\addons\main\script_macros.hpp"
+#include "\z\ace\addons\laserpointer\script_macros_config.hpp"
